refactor: optimizing faas cli and command-core
