+++ packages/command-core/package.json
@@ -7,7 +7,6 @@
     "deepmerge": "^4.2.2"
   },
   "devDependencies": {
-    "@types/mocha": "^5.2.5",
     "midway-bin": "1"
   },
   "files": [
@@ -18,9 +17,7 @@
     "build": "npm run lint && midway-bin build -c",
     "lint": "../../node_modules/.bin/tslint --format prose -c ../../tslint.json src/**/*.ts test/**/*.ts",
     "test": "npm run lint && NODE_ENV=test midway-bin test --ts --full-trace",
-    "debug": "npm run lint && NODE_ENV=test midway-bin test --ts --full-trace --inspect-brk=9229",
     "cov": "NODE_ENV=unittest midway-bin cov --ts",
-    "clean": "midway-bin clean",
     "autod": "midway-bin autod"
   },
   "gitHead": "b67e2753cbdcc91813067ba2a1bb1ce7e85a3dff"
