+++ packages/command-core/test/plugins/test.lg.ts
@@ -1,18 +1,18 @@
-import * as BasePlugin from '../../dist/plugin';
+import { BasePlugin } from '../../src';
 
 class LogPlugin extends BasePlugin {
-    provider = 'test';
-    commands = {
-        log: {
-            usage: 'log command',
-            lifecycleEvents: ['main']
-        }
-    };
-    hooks = {
-        'log:main': async () => {
-            await this.core.invoke(['invoke']);
-        },
-    };
+  provider = 'test';
+  commands = {
+    log: {
+      usage: 'log command',
+      lifecycleEvents: ['main'],
+    },
+  };
+  hooks = {
+    'log:main': async () => {
+      await this.core.invoke(['invoke']);
+    },
+  };
 }
 
 export default LogPlugin;
