+++ packages/faas-cli/bin/fun.js
@@ -1,5 +1,5 @@
 #!/usr/bin/env node
 'use strict';
-const CliClass = require('../dist');
-const cli = new CliClass.AliCli(process.argv);
+const { AliCLI } = require('../dist');
+const cli = new AliCLI(process.argv);
 cli.start();
