+++ packages/faas-cli/src/index.ts
@@ -1,11 +1,12 @@
-import { Cli } from '@midwayjs/faas-plugin-common';
+import { CLI } from '@midwayjs/faas-plugin-common';
 import { Invoke } from './invoke.plugin';
+
 export { Test } from '@midwayjs/faas-plugin-common';
 export { Invoke } from './invoke.plugin';
-export class AliCli extends Cli {
+
+export class AliCLI extends CLI {
   commands: any;
-  loadPlatformPlugin() {
-  }
+  loadPlatformPlugin() {}
 
   loadCommandInvoke() {
     this.core.addPlugin(Invoke);
